midterm/e-commerce/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-debug-toolbar==4.4.6
django-redis==5.4.0
djangorestframework==3.15.2
redis==5.1.1
sqlparse==0.5.1
